moves addBaseKPINode callback to BaseKPINode.tsx
