+++ src/components/ReactFlow/BaseKPINode.tsx
@@ -1,5 +1,5 @@
 // Define the custom BaseKPI node component with default node styling
-import { Handle, Position } from '@xyflow/react'
+import { Handle, Node, Position } from '@xyflow/react'
 import React from 'react'
 
 const BaseKPINode = ({ data }: { data: { label: string } }) => {
@@ -9,5 +9,18 @@ const BaseKPINode = ({ data }: { data: { label: string } }) => {
       <div>{data.label}</div>
     </div>
   )
+}
+// Define the function to add a new BaseKPI node to the flow
+export const addBaseKPINode = (
+  nodes: Node[],
+  setNodes: React.Dispatch<React.SetStateAction<Node[]>>,
+) => {
+  const newNode: Node = {
+    id: `basekpi_${nodes.length + 1}`,
+    position: { x: 0, y: 0 },
+    data: { label: 'BaseKPI' },
+    type: 'baseKpiNode',
+  }
+  setNodes((nds: Node[]) => nds.concat(newNode))
 }
 export default BaseKPINode
